api/pyproject.toml
[tool.poetry]
name = "API"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.1"
python-dotenv = "^0.17.1"
rich = "^10.2.2"
deta = "^0.8"
uvicorn = {extras = ["standard"], version = "^0.14.0"}

[tool.poetry.dev-dependencies]
black = "21.5b2"
flake8 = "^3.9.2"

[tool.poetry.scripts]
api = "main:start"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"